requirements.txt
apache-beam[gcp]==2.70.0
google-cloud-bigquery==3.40.1
google-cloud-pubsub==2.36.0
google-cloud-storage==2.19.0
faker==40.11.1
